Adding StarlingX_3rd.repo sourcelist for 3rd pkgs
For the mirror creation:
3rd packages in rpms_from_3rd_parties.lst file are downloaded
through urls with wget tool. Mostly times, some 3rd packages are missed
because the url download changes or the package is remove on the server.

Some servers that provide 3rd packages also works as Yum repository;
Taking advantage of this, Yumdownloader can handle the 3rd downloads
by adding these serves from rpms_from_3rd_parties.lst as Yum repos.

This have many advantages, for example:
if package A is download from server A is not available,
yumdownloader automatly can search the Package A the other servers.

In this way, the url is not required to download the 3rd package.
The file rpms_from_centos_3rd_parties.lst contains the name packages
that can be downloaded from Yum repos.

Change-Id: If1581b718e64b13ff6fd1484dd6af7b3889abb61
Signed-off-by: Felipe de Jesus Ruiz Garcia <felipe.de.jesus.ruiz.garcia@intel.com>