[WIP] StarlingX: Tar Compressed files repository
Primary intention for this review is to get community feedback.

The build of StarlingX relies, besides RPM Binaries and Sources, in this
repository which is a collection of packages in the form of Tar Compressed
files and 3 RPMs obtained from a Tar Compressed file.

This script and a text file containing a list of packages enable their download
and the creation of the repository based in common and specific requirements
dictated by the StarlingX building system recipes.

The source of the packages are:

- Tar Compressed files
  - Direct link: Getting them through a URL from a homepage
  - Git: Getting them through the Github API to download a specific commit / tag

The 2 categories we can divide the list of packages in the output directory:

- General hosted under "downloads" output directory.
- Puppet hosted under "downloads/puppet" output directory.

The file tarball-dl.lst text file contains 3 columns separated by a character "#":

- Column 1, name of package including extensions as it is referenced by the build
  system recipe, character "!" at the beginning of the name package denotes special
  handling is required.
- Column 2, name of the directory path after it is decompressed as it is referenced
  in the build system recipe.
- Column 3, the URL for the package download.

The file tarball-dl.sh script file will iterate over the tarball.lst text file and
execute specific tasks based on the name of the package:

- For the General category:
  - Most of the packages have a common process: download, decompressed, change the
    directory path and compressed.
  - We have 6 packages starting with the character "!" that require special handling
    besides the common process: remove directory, remove text from some files,
    clone a git repository, etc.
     - integrity-kmod-668a8270.tar.gz
     - mariadb-10.1.28.tar.gz
     - mvn.repo.tgz
     - qat1.7.upstream.l.1.0.3-42.tar.gz
     - tpm-kmod-668a8270.tar.gz
     - tss2-930.tar.gz
- For the Puppet category:
  - All packages have a common process: download, decompressed, change the directory
    path and compressed

Change-Id: I9e8ab59b20347ea33cf63308c213d4199be0e4fc
Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>
Co-authored-by: Mario Alfredo Carrillo Arevalo <mario.alfredo.c.arevalo@intel.com>