Add script to verify that repo's SRPMs matches with mirror's SRPMs
This DependenciesReviewer.py script verifies that SRPM's listed in
each stx- repo are listed in mirror's SRPMs lists.
This script attempts to identify broken dependencies before the
code is merged. It returns an exit code of 0 if it was sucessful,
and different to 0 if it was not.
It is planned that this will be executed by Zuul.

Story: 2002811
Change-Id: If78248a6f024086198714212108b64533d5f02b1
Signed-off-by: Marcela Rosales <marcela.a.rosales.jimenez@intel.com>