Add minimal comps.xml file into cgcs-centos generator
The comps.xml file provides the ability to install a group
of packages by default into mock, this enables the environment
with the minimal required packages. If this file is not present
mock won't be able to initialize and all packages compilation
will fail.

The common symptom of this problem is to see failures because
the adduser command cannot be found.

Story: 2002936
Task: 23033

Change-Id: I2e7d2fb86ff390cecdd53a0059b416dfffeb36ac
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>