Adding a tool to search and download missing packages automatically.
I would like to have the feedback from the community about this tool I made.

It is a script written in node.js that looks for a file on google and attempts
to download them, I know a lot of people won't want to install node on their
computers and that is why I created a Dockerfile so you can run this tool inside
a container.

The tool works this way:
- It receives a package name and the number of Google pages you want to search
on, it has a default of 10 pages in case you do not specify that parameter.
- Google API returns a set of URLs that contain the package name.
- The tool goes through all those pages looking for references of the package
inside the href tags of its html.
- Once it gets the URLs for every package it tries to download them inside of
a directory called downloads/<package_name>/

What still needs work?
- The packages that are downloaded might not be signed or have not integrity,
you will face this same issue even if you look for the packages manually around
the Internet.
- Maybe the script needs to receive a list of packages instead of just one at a
time, that might save some more time.

I would like to receive feedback from anybody that is willing to and maybe see
if it is worth keep working on this or not.

The instructions on how to use this tool are inside a README on the directory
stx-tools/centos-mirror-tools/package_auto_downloader/

Thanks

Signed-off-by: Guillermo Ponce <guillermo.a.ponce.castaneda@intel.com>

Change-Id: Ib7c62aeb704a8e1144061553a3750b41e478b68d