Update handling of mvn.repo.tgz
The special case handling of mvn.repo.tgz accesses loop variables
before they are set, resulting in the code using a value set during
the previous iteration. As a result, the script overwrites the
previously downloaded tarball with mvn.repo.tgz in the output
directory, and mvn.repo.tgz is missing.

This update moves the setting of the output path variables ahead of
the special case handlers.

Story: 2003058
Task: 23095

Change-Id: I3fbca2e056a584984d173942b0cbd46ec88579cc