Package Auto Downloader
This python script logs for a list of packages on Google and attempts to
download them.

The goal of the tool is to avoid the hassle of looking for packages that failed
during the mirror creation on Google, this tool will do it for you, so you won't
have to spend much time on it, it will print the URL where they are being
downloaded and will download as many as possible to ensure at least one of them
is a valid one.

The tool is written on python 3 and requires and has a requirements.txt file to
make it easier to configure and use, the use of a vurtual environment is adviced
but not required.

Please feel free to send me feedback on it.

Change-Id: I8d41e92c58d37c16e3ed380a59021ef0fb7669e6
Signed-off-by: Guillermo Ponce <guillermo.a.ponce.castaneda@intel.com>