Add mirror-check.sh to verify changes in upstream packages.
The mirror downloader uses a set of CentOS repositories to
get all the packages on specific versions, detailed on the .lst
files. If there are failures regarding new package versions we
can know until the end of the download process. This is something,
that can take a couple of hours.

This script uses the yum cache to verify if the package is
present in the current database. Therefore it is possible to
verify if the packages are available in the same version without
the need of download them. This slows down that time to verify
the mirror to about 5 minutes.

This script doesn't consider (yet) the 3rd party packages, those
needs to be verified in a different way.

Change-Id: I4472d9b6409c9498dcb749c9dee1b5ce43a06025
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>