Add designer workflow tool for downloading new packages and dependencies
This update introduces the starlingx_add_pkgs.sh utility. intended as
a designer workflow aid when adding new packages to the build or runtime
load. It uses the repo configuration files from centos-mirror-tools in
conjunction with the cgcs-centos-repo and build repos to determine
packages that must be download to satisfy requirements of new specified
packages being introduced. The downloaded RPMs are logged, the output
of which can be then used for populating the .lst files, as needed.

Change-Id: I655fab40e2b91dfeb9c5f714f8dc3f2be24cb8f4
Story: 2003294
Task: 24243
Signed-off-by: Don Penney <don.penney@windriver.com>