Work-in-progrees - remove root requirement for downloads
This work-in-progress is an update to some of the download scripts
which allow then to run in a non-containerized environment.  This
includes an environment where the user does not have sudo.

This may be the case, for example, where an automated process runs
the scripts outside of a container to update a local cache of
build-time requirements

Default behaviour continues to use sudo and run in a containerized
context, however additional options are introduced.  Depending on
how build system evolves, we may choose to make the default flow
of scripts not require/use sudo.

This change currently:
- provides for the option to not use "sudo" when executing commands
- allows for the usage of a custom yum.conf rather than use the
  system yum.conf and yum.repos.d
- allows for the user to not nessessarily change the group ID of
  downloaded artifacts
- adds an efficiency improvement when downloading tarballs

Change-Id: Ifcd764ca48ac5b8a40f90244dea2562e014e1dec
Signed-off-by: jmckenna <jason.mckenna@windriver.com>