Change centos 3rd parties list name and a condition in dl_rpms.sh
This in order to have the argument making true that condition (in
dl_rpms.sh) within the name of the list file (and later, in other
commit, think about removing $from paramenter in either
dowload_mirror.sh or dl_rpms.sh).

Also I adapted download_mirror.sh for this naming change, and the
log produced.

By standardizing the list files name it's possible to eliminate
one parameter in the download function in dl_rpms and reduce complexity.

Change-Id: I041fc9c704156215f06149e5b4c16cd92990e17c
Signed-off-by: Marcela Rosales <marcela.a.rosales.jimenez@intel.com>