Change lists name and dl_rpms.sh condition for downloading
This in order to have the argument making true that condition (in
dl_rpms.sh) within the name of the list file.

By standardizing the list files name it's possible to eliminate
one parameter in the download function in dl_rpms and reduce complexity.

I adapted download_mirror.sh for this naming change, and the
log produced.

Change-Id: I041fc9c704156215f06149e5b4c16cd92990e17c
Signed-off-by: Marcela Rosales <marcela.a.rosales.jimenez@intel.com>