Avoid downloading packages repeatedly
Due to the network or some other reason, we often need
to re-ran the "download_mirror.sh" script. In the process
of implementation, I found that some packages that have
already been downloaded will be re-downloaded. This takes
up bandwidth and increases deployment time. So I think we
should judge whether we have downloaded the package
before downloading it.

Change-Id: Id0d65c3b1d04fb80442e80040f5fbedbd8eb3abb