read releasever from yum.conf iff -c yum.conf is set
This solves a problem when the default releasever is not set for
distros that dont have a YUM as a standard tool.

Read from the yum.conf if set on the command line since that is mostly
likely going to be when we need to know this, in the default case (no -c
yum.conf) it will get the correct value from the system.

Change-Id: I21b0e8f881d054d8cdf6cdd831836a7f3e3020b2
Signed-off-by: Saul Wold <sgw@linux.intel.com>