Check for releasever in yum.conf before grepping
The dl_rpms.sh script may be passed an optional yum.conf file.  If
this file is used, then the releasever variable is read from the
rather than use the default releasever=7.  If the releasever is
not defined in the file, the grep which searches for the variable
will fail, and the dl_rpms.sh script exits with an error.

This change checks for the presence of releasever before attempting
to read it from the file.  Change made to dl_rpms.sh and
mirror-check.sh

Tested using yum.conf files both with and without releasever
present.

Change-Id: Ie13241e5662188d75b5d147537915c0a35ea7397
Closes-Bug: 1791113
Signed-off-by: Jason McKenna <jason.mckenna@windriver.com>