deployment: libvirt: update machine type and model
Tests should run on a comparably newer model of CPU and should be fixed
for all tests under virtualization.

Nehalem cpu model is recommended as one that has been used for this
purpose for a few years.

Specifying 'forbid' for fallback and 'exact' for mode will ensure that
hosts the Qemu emulator is up-to-date on the host, capable of emulating
the expect model.  The consistency will make test results reliable
across host machines.

The newer machine type supports ACPI for example.

The feature "vmx" is required on computes to support nesting.

Related-Bug: #1790716

Change-Id: Ib338501b4805532a9d1ebaf55735f9084bfa8e96
Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>