deployment: libvirt: correct internal/external net
The INTERNAL and EXTERNAL networks configured by destroy and setup
network scripts are inverted with reference to the default
config_controller selections as well as the wiki
StarlingX/Installation_Guide_Virtual_Environment/Controller_Storage.

The most pressing concern in this is that what was actually the internal
management network was being nat'd.

Also, under normal testing it is not recommended to address the internal
management network from the virtualization host.  There are unlikely to be
configurations that will have external devices addressing the management
network.

Delete what was "EXTERNAL", actually internal, and rename "INTERNAL" to
be EXTERNAL.

Related-Bug: #1790716

Change-Id: I08f9e7712fed120001c864903880e13e9478764d
Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>