Add unified dockerfile
The objective of this commit is to unify the customization of
the follow docker images in one file.
- Dockerfile.centos73
- Dockerfile.centos73.TC-builder
- centos-mirror-tools/Dockerfile

This is order to create a generic StarlingX docker image, this image
should be able to handle all building process.

Furthermore this unification will reduce the instructions of the
installation guide for our end users.

Story: 2003712
Task: 226370

Change-Id: Id9b405743763be2e0d7aa37c8070c6788bc099c3
Signed-off-by: Mario Alfredo Carrillo Arevalo <mario.alfredo.c.arevalo@intel.com>