Changing lshell from centos to centos3rdparty list
Right now lshell is looked for in StarlingX_3rd_epel_blizoo_mk.repo,
and it should be in centos3rdparties instead of centos list.

As it is now, there's a bug because right before centos list is
downloaded, the downwload_mirror.sh script removes yum 3rd parties config
files, and later the yumdownloader doesn't find the lshell package.

Change-Id: I1eab20226fb92b4b2e05f50958112092bb643bc1
Signed-off-by: Marcela Rosales <marcela.a.rosales.jimenez@intel.com>