Add -t to branch-stx.sh to enable a tag-only mode
We need to be able to tag final releases without creating a new branch.
Changing the starting branch (formerly 'master') to be the same as the
destination branch has the desired effect so the -t option does that
in addition to skipping the automatic .gitreview update.

Note that setting the source branch (SRC_BRANCH) to something other than
master is also something we will need in the regualt branching mode
now that we have default (working) branches in some of the staging
repos set to names similar to stx/v1.2.3 to represent the starting point
in the upstream repo of our working branch.

Change-Id: Id0cb485a4f8a2ebd3feaf602a140f66ed3b91322
Signed-off-by: Dean Troyer <dtroyer@gmail.com>