Move docker creation and deletion to tb.sh
A makefile was used only for creation and clean up of the docker
image for building. In this process, all the environment variables
were sourced and included into the Makefile. However this can fail
on systems that has functions and cause a failing makefile.

Trying to fixing this issue I realize that this Makefile ain't needed,
as all the logic for docker management is under `tb.sh`, therefore makes
sense to remove this Makefile and use this script for the entire process.

Story: 2002560

Change-Id: I2795593521a5d16dc2033861946fd09c2270424e
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>