download_mirrors.sh: cengn urls are not validated before substitution
Problem
=======
download_mirrors.sh by default substitutes urls with their cengn
equivalent. If the user is trying to introduce a new repo, a CENGN
mirror of the repo will not yet exist. The substituted url will be added
to the yum configuration despite being invalid. All subsequent
yumdownload attempts will fail with a 404 on the repodata of the
non-existent icengn repo.

Solution
========
   Only substitute yum repo urls with the cengn equivalent if
cengn actually has the repo.

Closes-bug: 1824877
Change-Id: Ifa262212d67e096cc29131e5738aec0365ed9893
Signed-off-by: Scott Little <scott.little@windriver.com>