Avoid to produce broken tarballs by checking the mime type.
The tarball downloader uses wget to retrieve content from the urls.
In some cases, this url doesn't point directly to a file in the server.
For cases like sourceforge, the url points to an API that will deliver
the requested file. This is problematic for wget because still gets a
200 HTTP reponse even if the is not there, and then downloads a text
file with the content of the requested page.

This fix proposes to check the mime type of the downloaded file and if
the file is not one of the used types the it will delete it.

Closes-Bug: 1795739

Change-Id: Ia9286ca094389d8d5196fce8ccda8a38cc754075
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>